[BUGFIX release] Refactor Model.reopen to use mixins
This fixes bugs where users import Model form 'ember-data/model' and it was missing required methods
